Web App/api/requirements.txt
nltk==3.5
Flask_RESTful==0.3.8
numpy==1.19.2
pandas==1.2.3
Flask_Cors==3.0.10
tensorflow-cpu
tweepy==3.10.0
Flask==1.1.2
gunicorn